# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Restful Provisioning Protocol (RPP) ",
    "description": "A Modern Domain Registry API",
    "version": "0.0.1"
  },
  "paths": {
...
    "/hosts/": {
      "post": {
        "tags": [
          "Hosts"
        ],
        "summary": "Create Host",
        "operationId": "do_create_hosts__post",
        "security": [
          {
            "HTTPBasic": []
          },
          {
            "HTTPBasic": []
          }
        ],
        "parameters": [
          {
            "name": "rpp-cl-trid",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Rpp-Cl-Trid"
            }
          },
          {
            "name": "session_id",
            "in": "cookie",
            "required": false,
            "schema": {
              "type": "string",
              "title": "Session Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/HostCreateRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Host created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BaseResponseModel"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/hosts/{host}": {
      "get": {
        "tags": [
          "Hosts"
        ],
        "summary": "Get Host Info",
        "operationId": "do_info_hosts__host__get",
        "security": [
          {
            "HTTPBasic": []
          },
          {
            "HTTPBasic": []
          }
        ],
        "parameters": [
          {
            "name": "host",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Host"
            }
          },
          {
            "name": "rpp-cl-trid",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Rpp-Cl-Trid"
            }
          },
          {
            "name": "session_id",
            "in": "cookie",
            "required": false,
            "schema": {
              "type": "string",
              "title": "Session Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BaseResponseModel"
                }
              }
            },
            "headers": {
              "RPP-Code": {
                "description": "RPP result code",
                "schema": {
                  "type": "string"
                }
              },
              "RPP-Cltrid": {
                "description": "RPP client transaction ID",
                "schema": {
                  "type": "string"
                }
              },
              "RPP-Svtrid": {
                "description": "RPP server transaction ID",
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Hosts"
        ],
        "summary": "Delete Host",
        "operationId": "do_delete_hosts__host__delete",
        "security": [
          {
            "HTTPBasic": []
          },
          {
            "HTTPBasic": []
          }
        ],
        "parameters": [
          {
            "name": "host",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Host"
            }
          },
          {
            "name": "rpp-cl-trid",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Rpp-Cl-Trid"
            }
          },
          {
            "name": "session_id",
            "in": "cookie",
            "required": false,
            "schema": {
              "type": "string",
              "title": "Session Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response",
            "headers": {
              "RPP-Code": {
                "description": "RPP result code",
                "schema": {
                  "type": "string"
                }
              },
              "RPP-Cltrid": {
                "description": "RPP client transaction ID",
                "schema": {
                  "type": "string"
                }
              },
              "RPP-Svtrid": {
                "description": "RPP server transaction ID",
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "Hosts"
        ],
        "summary": "Update Host",
        "operationId": "do_update_hosts__host__patch",
        "security": [
          {
            "HTTPBasic": []
          },
          {
            "HTTPBasic": []
          }
        ],
        "parameters": [
          {
            "name": "host",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Host"
            }
          },
          {
            "name": "rpp-cl-trid",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Rpp-Cl-Trid"
            }
          },
          {
            "name": "session_id",
            "in": "cookie",
            "required": false,
            "schema": {
              "type": "string",
              "title": "Session Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/HostUpdateRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BaseResponseModel"
                }
              }
            },
            "headers": {
              "RPP-Code": {
                "description": "RPP result code",
                "schema": {
                  "type": "string"
                }
              },
              "RPP-Cltrid": {
                "description": "RPP client transaction ID",
                "schema": {
                  "type": "string"
                }
              },
              "RPP-Svtrid": {
                "description": "RPP server transaction ID",
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/hosts/{host}/processes/{proc_name}/{proc_id}": {
      "get": {
        "tags": [
          "Hosts"
        ],
        "summary": "List Processes",
        "operationId": "do_list_processes_hosts__host__processes__proc_name___proc_id__get",
        "security": [
          {
            "HTTPBasic": []
          },
          {
            "HTTPBasic": []
          }
        ],
        "parameters": [
          {
            "name": "host",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Host"
            }
          },
          {
            "name": "proc_name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Proc Name"
            }
          },
          {
            "name": "proc_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Proc Id"
            }
          },
          {
            "name": "rpp-cl-trid",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Rpp-Cl-Trid"
            }
          },
          {
            "name": "rpp-authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Rpp-Authorization"
            }
          },
          {
            "name": "session_id",
            "in": "cookie",
            "required": false,
            "schema": {
              "type": "string",
              "title": "Session Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "headers": {
              "RPP-Code": {
                "description": "RPP result code",
                "schema": {
                  "type": "string"
                }
              },
              "RPP-Cltrid": {
                "description": "RPP client transaction ID",
                "schema": {
                  "type": "string"
                }
              },
              "RPP-Svtrid": {
                "description": "RPP server transaction ID",
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/messages/": {
      "get": {
        "tags": [
          "Messages"
        ],
        "summary": "Get Messages",
        "operationId": "do_get_messages_messages__get",
        "security": [
          {
            "HTTPBasic": []
          },
          {
            "HTTPBasic": []
          }
        ],
        "parameters": [
          {
            "name": "rpp-cl-trid",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Rpp-Cl-Trid"
            }
          },
          {
            "name": "session_id",
            "in": "cookie",
            "required": false,
            "schema": {
              "type": "string",
              "title": "Session Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BaseResponseModel"
                }
              }
            },
            "headers": {
              "RPP-Code": {
                "description": "RPP result code",
                "schema": {
                  "type": "string"
                }
              },
              "RPP-Cltrid": {
                "description": "RPP client transaction ID",
                "schema": {
                  "type": "string"
                }
              },
              "RPP-Svtrid": {
                "description": "RPP server transaction ID",
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/messages/{message_id}": {
      "delete": {
        "tags": [
          "Messages"
        ],
        "summary": "Delete Message",
        "operationId": "do_delete_message_messages__message_id__delete",
        "security": [
          {
            "HTTPBasic": []
          },
          {
            "HTTPBasic": []
          }
        ],
        "parameters": [
          {
            "name": "message_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Message Id"
            }
          },
          {
            "name": "rpp-cl-trid",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Rpp-Cl-Trid"
            }
          },
          {
            "name": "session_id",
            "in": "cookie",
            "required": false,
            "schema": {
              "type": "string",
              "title": "Session Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response",
            "headers": {
              "RPP-Code": {
                "description": "RPP result code",
                "schema": {
                  "type": "string"
                }
              },
              "RPP-Cltrid": {
                "description": "RPP client transaction ID",
                "schema": {
                  "type": "string"
                }
              },
              "RPP-Svtrid": {
                "description": "RPP server transaction ID",
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "BaseResponseModel": {
        "properties": {
          "@type": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "@Type"
          },
          "result": {
            "items": {
              "$ref": "#/components/schemas/ResultModel"
            },
            "type": "array",
            "title": "Result"
          },
          "messages": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/MessageQueueModel"
              },
              {
                "type": "null"
              }
            ]
          },
          "resData": {
            "anyOf": [
              {},
              {
                "type": "null"
              }
            ],
            "title": "Resdata"
          },
          "extension": {
            "anyOf": [
              {},
              {
                "type": "null"
              }
            ],
            "title": "Extension"
          }
        },
        "type": "object",
        "required": [
          "result"
        ],
        "title": "BaseResponseModel"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "HostAddr": {
        "properties": {
          "v4": {
            "anyOf": [
              {
                "items": {
                  "type": "string",
                  "format": "ipvanyaddress"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "V4",
            "default": []
          },
          "v6": {
            "anyOf": [
              {
                "items": {
                  "type": "string",
                  "format": "ipvanyaddress"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "V6",
            "default": []
          }
        },
        "type": "object",
        "title": "HostAddr"
      },
      "HostCreateRequest": {
        "properties": {
          "@type": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "@Type"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "addr": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/HostAddr"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "required": [
          "name"
        ],
        "title": "HostCreateRequest"
      },
      "HostUpdateAddOrRemove": {
        "properties": {
          "addr": {
            "items": {
              "$ref": "#/components/schemas/HostAddr"
            },
            "type": "array",
            "title": "Addr"
          },
          "status": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/StatusModel"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Status"
          }
        },
        "type": "object",
        "required": [
          "addr"
        ],
        "title": "HostUpdateAddOrRemove"
      },
      "HostUpdateChange": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          }
        },
        "type": "object",
        "required": [
          "name"
        ],
        "title": "HostUpdateChange"
      },
      "HostUpdateRequest": {
        "properties": {
          "@type": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "@Type"
          },
          "add": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/HostUpdateAddOrRemove"
              },
              {
                "type": "null"
              }
            ]
          },
          "remove": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/HostUpdateAddOrRemove"
              },
              {
                "type": "null"
              }
            ]
          },
          "change": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/HostUpdateChange"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "title": "HostUpdateRequest"
      },
...
      "MessageQueueModel": {
        "properties": {
          "count": {
            "type": "integer",
            "title": "Count"
          },
          "id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Id"
          },
          "qDate": {
            "type": "string",
            "format": "date-time",
            "title": "Qdate"
          },
          "message": {
            "type": "string",
            "title": "Message"
          }
        },
        "type": "object",
        "required": [
          "count"
        ],
        "title": "MessageQueueModel"
      },
...
      "ResultModel": {
        "properties": {
          "code": {
            "type": "integer",
            "title": "Code"
          },
          "message": {
            "type": "string",
            "title": "Message"
          },
          "lang": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Lang"
          },
          "value": {
            "anyOf": [
              {
                "items": {},
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Value"
          },
          "extValue": {
            "anyOf": [
              {
                "items": {},
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Extvalue"
          }
        },
        "type": "object",
        "required": [
          "code",
          "message"
        ],
        "title": "ResultModel"
      },
...
      "StatusModel": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reason"
          }
        },
        "type": "object",
        "required": [
          "name"
        ],
        "title": "StatusModel"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "HTTPBasic": {
        "type": "http",
        "scheme": "basic"
      }
    }
  }
}